console: add tabs component
## Description 🔖

[VIDEO](https://www.loom.com/share/5844476e527f4e1d9ec296bc3a4bb932)

This PR creates a Tabs component that wraps the Radix UI component.

## Solution and Design 🎨

The component accepts all the props of the Radix UI tab root component and a list of items that specify the label and the content of the tabs. The component can be used in a controlled way (e.g., the state of the tab is stored in the URL) or uncontrolled.

For styling radix ui states I've added this library: https://github.com/ecklf/tailwindcss-radix

## Review Setup 💻

Run storybook and look for Tabs component

## Review checklist. 📋

- [ ] stories work
- [ ] doc is clear

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/5667
GitOrigin-RevId: 0e73ef1026d6673c3c2dd413b986fa8395abb1f6