server/tests-py: If a port is busy, try to figure out why.
Every now and again, I see in CI that a test will abort early because port 8080 is busy. Given that we're running in a fresh Docker container, this makes no sense to me.

On failure, it will now run `lsof -i :$PORT` to try and figure out what's listening on that port.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/6285
GitOrigin-RevId: 5eb58d016b2f43872e2c78a7d89d27cfc3e22312