CI: Disable Python integration test groups covered by the new job.
We now run a lot of these at once by spawning a new HGE instance per
test class. The commented-out tests are now duplicated.

We have a lot of work to do to take care of the rest, but this should be
a decent start. With this change, we drop from 42 to 19 test groups.

[NDAT-533]: https://hasurahq.atlassian.net/browse/NDAT-533?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/8060
GitOrigin-RevId: e38e1a72cfa6eafac97065a435d2c8e1508b834e